requirements.txt
opencv-python>=4.8.0
openai>=1.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
mediapipe>=0.10.0
pillow>=10.0.0
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'
